Change package structure and java version
