--- src/main/java/io/swagger/api/ApiException.java
@@ -1,11 +1,0 @@
-package io.swagger.api;
-
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.SpringCodegen", date = "2016-11-25T08:23:20.639Z")
-
-public class ApiException extends Exception{
-	private int code;
-	public ApiException (int code, String msg) {
-		super(msg);
-		this.code = code;
-	}
-}
