--- src/main/java/io/swagger/api/NotFoundException.java
@@ -1,11 +1,0 @@
-package io.swagger.api;
-
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.SpringCodegen", date = "2016-11-25T08:23:20.639Z")
-
-public class NotFoundException extends ApiException {
-	private int code;
-	public NotFoundException (int code, String msg) {
-		super(code, msg);
-		this.code = code;
-	}
-}
