+++ src/main/java/uk/ac/sanger/scgcf/barcodegenerator/RFC3339DateFormat.java
@@ -1,4 +1,4 @@
-package io.swagger;
+package uk.ac.sanger.scgcf.barcodegenerator;
 
 import com.fasterxml.jackson.databind.util.ISO8601DateFormat;
 import com.fasterxml.jackson.databind.util.ISO8601Utils;
