+++ src/main/java/uk/ac/sanger/scgcf/barcodegenerator/api/ApiOriginFilter.java
@@ -1,11 +1,11 @@
-package io.swagger.api;
+package uk.ac.sanger.scgcf.barcodegenerator.api;
 
 import java.io.IOException;
 
 import javax.servlet.*;
 import javax.servlet.http.HttpServletResponse;
 
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.SpringCodegen", date = "2016-11-25T08:23:20.639Z")
+@javax.annotation.Generated(value = "class uk.ac.sanger.scgcf.barcodegenerator.codegen.languages.SpringCodegen", date = "2016-11-25T08:23:20.639Z")
 
 public class ApiOriginFilter implements javax.servlet.Filter {
 	@Override
