+++ src/main/java/uk/ac/sanger/scgcf/barcodegenerator/api/BarcodesApi.java
@@ -1,23 +1,22 @@
-package io.swagger.api;
-
-import io.swagger.model.Barcode;
-import io.swagger.model.Error;
-import io.swagger.model.SingleBarcodePayload;
-
-import io.swagger.annotations.*;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RequestPart;
-import org.springframework.web.multipart.MultipartFile;
+package uk.ac.sanger.scgcf.barcodegenerator.api;
 
 import java.util.List;
 
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.SpringCodegen", date = "2016-11-25T08:23:20.639Z")
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import uk.ac.sanger.scgcf.barcodegenerator.persistence.model.Barcode;
+import uk.ac.sanger.scgcf.barcodegenerator.persistence.model.SingleBarcodePayload;
+
+@javax.annotation.Generated(value = "class uk.ac.sanger.scgcf.barcodegenerator.codegen.languages.SpringCodegen", date = "2016-11-25T08:23:20.639Z")
 
 @Api(value = "barcodes", description = "the barcodes API")
 public interface BarcodesApi {
