+++ src/main/java/uk/ac/sanger/scgcf/barcodegenerator/api/BatchBarcodesApiController.java
@@ -1,10 +1,9 @@
-package io.swagger.api;
-
-import io.swagger.model.BarcodesPayload;
-import io.swagger.model.BatchBarcodes;
-import io.swagger.model.Error;
+package uk.ac.sanger.scgcf.barcodegenerator.api;
 
 import io.swagger.annotations.*;
+import uk.ac.sanger.scgcf.barcodegenerator.persistence.model.BarcodesPayload;
+import uk.ac.sanger.scgcf.barcodegenerator.persistence.model.BatchBarcodes;
+import uk.ac.sanger.scgcf.barcodegenerator.persistence.model.Error;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,7 +18,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
 
 
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.SpringCodegen", date = "2016-11-25T08:23:20.639Z")
+@javax.annotation.Generated(value = "class uk.ac.sanger.scgcf.barcodegenerator.codegen.languages.SpringCodegen", date = "2016-11-25T08:23:20.639Z")
 
 @Controller
 public class BatchBarcodesApiController implements BatchBarcodesApi {
