+++ src/main/java/uk/ac/sanger/scgcf/barcodegenerator/api/NotFoundException.java
@@ -1,0 +1,11 @@
+package uk.ac.sanger.scgcf.barcodegenerator.api;
+
+@javax.annotation.Generated(value = "class uk.ac.sanger.scgcf.barcodegenerator.codegen.languages.SpringCodegen", date = "2016-11-25T08:23:20.639Z")
+
+public class NotFoundException extends ApiException {
+	private int code;
+	public NotFoundException (int code, String msg) {
+		super(code, msg);
+		this.code = code;
+	}
+}
