+++ src/main/java/uk/ac/sanger/scgcf/barcodegenerator/configuration/HomeController.java
@@ -1,4 +1,4 @@
-package io.swagger.configuration;
+package uk.ac.sanger.scgcf.barcodegenerator.configuration;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
