+++ src/main/resources/application.properties
@@ -1,5 +1,43 @@
 springfox.documentation.swagger.v2.path=/api-docs
 server.contextPath=
 server.port=8080
-spring.jackson.date-format=io.swagger.RFC3339DateFormat
+spring.jackson.date-format=uk.ac.sanger.scgcf.barcodegenerator.RFC3339DateFormat
 spring.jackson.serialization.WRITE_DATES_AS_TIMESTAMPS=false
+
+# ===============================
+# = DATA SOURCE
+# ===============================
+
+# Set here configurations for the database connection
+
+# Connection url for the database "netgloo_blog"
+spring.datasource.url = jdbc:mysql://localhost:3306/barcode_generator
+
+# Username and password
+spring.datasource.username = root
+spring.datasource.password = 
+
+# Keep the connection alive if idle for a long time (needed in production)
+spring.datasource.testWhileIdle = true
+spring.datasource.validationQuery = SELECT 1
+
+# ===============================
+# = JPA / HIBERNATE
+# ===============================
+
+# Use spring.jpa.properties.* for Hibernate native properties (the prefix is
+# stripped before adding them to the entity manager).
+
+# Show or not log for each sql query
+spring.jpa.show-sql = true
+
+# Hibernate ddl auto (create, create-drop, update): with "update" the database
+# schema will be automatically updated accordingly to java entities found in
+# the project
+spring.jpa.hibernate.ddl-auto = update
+
+# Naming strategy
+spring.jpa.hibernate.naming-strategy = org.hibernate.cfg.ImprovedNamingStrategy
+
+# Allows Hibernate to generate SQL optimized for a particular DBMS
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
